Add session.write, session.new_tab APIs and tmux sessions example plugin
New Lua/native plugin APIs:
- session.write(text): write bytes to the focused terminal's PTY
- session.new_tab(command?, plain?): open a new tab with optional
  command and plain shell mode (bypasses terminal.shell config)

Infrastructure:
- HostApi: add write_to_pty and new_tab FFI functions
- Bridge: PENDING_PTY_WRITES and PENDING_NEW_TABS queues drained by
  the app render loop
- sessions.rs: create_plain_session() ignores terminal.shell config
- Fix u8 overflow in tab_bar hover color averaging (light mode crash)

Example plugin:
- tmux-sessions.lua: hierarchical tree of tmux sessions with windows,
  right-click rename/kill/detach, cleanup button, attach via new tab

+++ crates/conch_app/src/tab_bar.rs
@@ -268,9 +268,9 @@ impl TabBarColors {
             bar_bg: theme.surface,
             active_bg: theme.surface_raised,
             hover_bg: Color32::from_rgb(
-                (theme.surface.r() + theme.surface_raised.r()) / 2,
-                (theme.surface.g() + theme.surface_raised.g()) / 2,
-                (theme.surface.b() + theme.surface_raised.b()) / 2,
+                ((theme.surface.r() as u16 + theme.surface_raised.r() as u16) / 2) as u8,
+                ((theme.surface.g() as u16 + theme.surface_raised.g() as u16) / 2) as u8,
+                ((theme.surface.b() as u16 + theme.surface_raised.b() as u16) / 2) as u8,
             ),
             text: theme.text,
             text_dim: theme.text_secondary,
